+++ electron/match/matching.js
@@ -1,28 +1,12 @@
 var fs = require('fs');
 var natural = require('natural');
-var regMatch = require('./regMatch');
 var formatVariable = require('./formatVariable');
+var testPhrases = require('./testers/testPhrases');
+var getMatchByScore = require('./testers/getMatchByScore');
+var phoneticsTest = require('./testers/phoneticsTest');
+var JWDTest = require('./testers/JWDTest');
 
-var JaroWinklerDistance = natural.JaroWinklerDistance;
-var Metaphone = natural.Metaphone;
-var soundEx = natural.SoundEx;
-
-soundEx.attach();
-
-module.exports.testPhrases = function (phrases, _actionPrefix) {
-  return regMatch(phrases, _actionPrefix);
-};
-module.exports.JWDTest = function (key, _actionPrefix) {
-  return JaroWinklerDistance(key, _actionPrefix);
-};
-module.exports.phoneticsTest = function (actionPrefix, key) {
-  console.log(Metaphone.process(key));
-  console.log(Metaphone.process(actionPrefix));
-  return module.exports.JWDTest(Metaphone.process(key), Metaphone.process(actionPrefix));
-};
-
-
-var matching = function (actionPrefix, variable, commandsObj) {
+module.exports = function (actionPrefix, variable, commandsObj) {
   var _actionPrefix = actionPrefix.toLowerCase();
   var actionObj = {};
   actionObj.exact = false;
@@ -34,7 +18,7 @@ var matching = function (actionPrefix, variable, commandsObj) {
   var closeMatchThreshold = 0.6;
 
   var phrases = commandsObj.phrases;
-  var actions = commandsObj.rawCommands; // TODO: change to rawCommands
+  var actions = commandsObj.rawCommands;
   var argCommands = commandsObj.parsedCommands.argCommands;
   var exactCommands = commandsObj.parsedCommands.exactCommands;
 
@@ -49,44 +33,25 @@ var matching = function (actionPrefix, variable, commandsObj) {
     return actionObj;
   }
 
-
-  for (var key in phrases) {
-
-    if (module.exports.testPhrases(phrases[key], _actionPrefix)) {
-      console.log('added phrase found');
-      actionObj.exact = true;
-      if (variable && argCommands[_actionPrefix]) {
-        actionObj.action = formatVariable(argCommands[key], variable, commandsObj);
-      } else {
-        actionObj.action = exactCommands[key];
-      }
-      return actionObj;
+  var addedPhraseTest = testPhrases(phrases, _actionPrefix);
+  if (addedPhraseTest) {
+    console.log('added phrase found: ', addedPhraseTest);
+    actionObj.exact = true;
+    if (variable && argCommands[addedPhraseTest]) {
+      actionObj.action = formatVariable(argCommands[addedPhraseTest], variable, commandsObj);
+    } else {
+      actionObj.action = exactCommands[addedPhraseTest];
     }
-
-    if (module.exports.JWDTest(_actionPrefix, key) > exactMatchThreshold) {
-      actionObj.exact = false;
-      if (variable && argCommands[key]) {
-        actionObj.action = formatVariable(argCommands[key], variable, commandsObj);
-      } else {
-        actionObj.action = exactCommands[key];
-      }
-      return actionObj;
-    }
-
-    if (module.exports.phoneticsTest(_actionPrefix, key) > closeMatchThreshold) {
-      console.log('phonetic match found');
-      actionObj.exact = false;
-      actionObj.guessedCommand = key;
-      if (variable && argCommands[key]) {
-        actionObj.action = formatVariable(argCommands[key], variable, commandsObj);
-      } else {
-        actionObj.action = exactCommands[key];
-      }
-      return actionObj;
+  } else {
+    var key = getMatchByScore(Object.keys(actions), _actionPrefix);
+    console.log('guessing ' + key);
+    actionObj.exact = false;
+    actionObj.guessedCommand = key;
+    if (variable && argCommands[key]) {
+      actionObj.action = formatVariable(argCommands[key], variable, commandsObj);
+    } else {
+      actionObj.action = exactCommands[key];
     }
   }
-
   return actionObj;
 };
-
-module.exports.matching = matching;
