+++ electron/match/phrases.json
@@ -1,1 +1,1 @@
-{"kyle cho pro tip":["kyle cho pro tip","say hello","kay","kyoto proto","paltrow protest","how to protect","kyosho concept","kyoto protocol","icago protest","house of cards","kyosho protech","how to i fratelli","kyoto protest","township road 7","osmoprof 10","tow truck protest","kyle show protest"],"check the":["check the"],"dim screen":["dim screen","controversy"],"open":["open"],"list folders":["list folders"],"volume down":["volume down","phone down"],"volume up":["volume up"],"volume off":["volume off"],"Shut up":["Shut up"],"Play":["Play"],"Youtube":["Youtube"],"shutdown":["shutdown"],"google":["google"],"Wikipedia":["Wikipedia"],"null":["null"],"say hello":["say hello"],"enhance":["enhance"],"zoom out":["zoom out"],"dance for me":["dance for me"],"new command":["new command"],"old command":["old command","old command"],"jarvis":["jarvis"],"do awesome things":["do awesome things"],"Free Willy":["Free Willy"],"shut up":"shut up","play":"play","youtube":"youtube","wikipedia":"wikipedia","dehance":"dehance","free willy":"free willy"}
+{"kyle cho pro tip":["kyle cho pro tip","say hello","kay","kyoto proto","paltrow protest","how to protect","kyosho concept","kyoto protocol","icago protest","house of cards","kyosho protech","how to i fratelli","kyoto protest","township road 7","osmoprof 10","tow truck protest","kyle show protest"],"check the":["check the"],"dim screen":["dim screen","controversy"],"open":["open"],"list folders":["list folders"],"volume down":["volume down","phone down"],"volume up":["volume up"],"volume off":["volume off"],"Shut up":["Shut up"],"Play":["Play"],"Youtube":["Youtube"],"shutdown":["shutdown"],"google":["google"],"Wikipedia":["Wikipedia"],"null":["null"],"say hello":["say hello"],"enhance":["enhance","manhattan"],"zoom out":["zoom out","shoe mountain"],"dance for me":["dance for me"],"new command":["new command"],"old command":["old command","old command"],"jarvis":["jarvis"],"do awesome things":["do awesome things"],"Free Willy":["Free Willy"],"shut up":"shut up","play":"play","youtube":"youtube","wikipedia":"wikipedia","dehance":"dehance","free willy":"free willy"}
