+++ electron/test/parsing.spec.js
@@ -1,0 +1,7 @@
+var expect = require('chai').expect;
+
+describe('Should parse commands', function (done) {
+  it('should match strings that are arguments', function (done) {
+    done();
+  });
+});
